Make common popup for file and folder
