+++ src/components/folder-card.js
@@ -10,40 +10,9 @@ import { setActiveItems } from '../actions/itemActions'
 import { deleteFolder } from '../actions/folderActions'
 import { ITEM_TYPE } from '../constants'
 
-const options = [
-  { key: '1', label: 'Edit', icon: 'edit' },
-  { key: '2', label: 'Download' },
-  { key: '3', label: 'Delete', icon: 'delete' }
-]
-
-function createContextFromEvent (e) {
-  const left = e.clientX
-  const top = e.clientY
-  const right = left + 1
-  const bottom = top + 1
-
-  return {
-    getBoundingClientRect: () => ({
-      left,
-      top,
-      right,
-      bottom,
-
-      height: 0,
-      width: 0
-    })
-  }
-}
-
 class FolderCard extends Component {
-  constructor (props) {
+  constructor(props) {
     super(props)
-    this.state = {
-      showEditModal: false,
-      showDeleteModal: false,
-      isPopupOpen: false
-    }
-    this.contextRef = React.createRef()
     this.ref = React.createRef()
   }
 
@@ -61,27 +30,7 @@ class FolderCard extends Component {
     }
   }
 
-  handleOptions = {
-    1: () => {
-      this.setState({ showEditModal: true })
-    },
-    2: () => {},
-    3: () => {
-      this.setState({ showDeleteModal: true })
-    }
-  }
-
-  handleDelete = () => {
-    const { deleteFolder, activeItems, setActiveItems } = this.props
-    if (activeItems.length === 1 && activeItems[0].type === ITEM_TYPE.folder) {
-      deleteFolder(activeItems[0].obj.id)
-    }
-    setActiveItems([])
-    this.setState({ showDeleteModal: false })
-  }
-
-  render () {
-    const { isPopupOpen, showDeleteModal, showEditModal } = this.state
+  render() {
     const { index, folder, activeItems } = this.props
     const folderName =
       folder.folderName.length > 12
@@ -90,12 +39,9 @@ class FolderCard extends Component {
     return (
       <div
         id={`grid-card-${index}`}
-        styleName='grid.file-card'
+        styleName="grid.file-card"
         onContextMenu={e => {
           e.preventDefault()
-          console.log(e)
-          this.contextRef = createContextFromEvent(e)
-          this.setState({ isPopupOpen: true })
           this.handleSelect(e)
         }}
         secondary
@@ -109,10 +55,10 @@ class FolderCard extends Component {
               : 'grid.folder-inactive'
           }
         >
-          <div styleName='grid.folder-icon' onClick={e => this.handleSelect(e)}>
+          <div styleName="grid.folder-icon" onClick={e => this.handleSelect(e)}>
             <Icon
-              size='huge'
-              color='grey'
+              size="huge"
+              color="grey"
               name={'folder open'}
               onDoubleClick={() => {
                 const url = `/file-manager/${this.props.match.params.filemanager}/${folder.id}/`
@@ -120,7 +66,7 @@ class FolderCard extends Component {
               }}
             />
           </div>
-          <div styleName='grid.folder-name'>
+          <div styleName="grid.folder-name">
             <p
               onDoubleClick={() => {
                 const url = `/file-manager/${this.props.match.params.filemanager}/${folder.id}/`
@@ -131,37 +77,6 @@ class FolderCard extends Component {
             </p>
           </div>
         </div>
-        <Popup
-          basic
-          context={this.contextRef}
-          onClose={() => this.setState({ isPopupOpen: false })}
-          open={isPopupOpen}
-        >
-          <Menu vertical>
-            {options.map(option => (
-              <Menu.Item
-                key={option.key}
-                name={option.label}
-                icon={option.icon ? option.icon : false}
-                onClick={() => {
-                  this.setState({ isPopupOpen: false })
-                  this.handleOptions[option.key]()
-                }}
-              />
-            ))}
-          </Menu>
-        </Popup>
-        <ConfirmModal
-          show={showDeleteModal}
-          handleClose={() => {
-            this.setState({ showDeleteModal: false })
-          }}
-          handleSubmit={() => {
-            this.handleDelete()
-          }}
-          type='remove'
-          item='file'
-        />
       </div>
     )
   }
@@ -171,7 +86,7 @@ const mapStateToProps = state => {
   return {
     gridViewActiveIndex: state.files.gridViewActiveIndex,
     activeFolder: state.folders.activeFolder,
-    activeItems: state.items.activeItems
+    activeItems: state.items.activeItems,
   }
 }
 
@@ -182,7 +97,7 @@ const mapDispatchToProps = dispatch => {
     },
     deleteFolder: id => {
       dispatch(deleteFolder(id))
-    }
+    },
   }
 }
